+++ src/components/Course/Form/store.ts
@@ -1,8 +1,7 @@
-import { useState, useEffect, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useNavigate } from 'react-router-dom';
 import * as Yup from 'yup';
-import { useForm, Controller } from 'react-hook-form';
+import { useForm } from 'react-hook-form';
 import { yupResolver } from '@hookform/resolvers/yup';
 import request from 'src/hooks/useRequest';
 import { useCategory } from 'src/hooks/useCategory';
